feat: add the automatic upgrading capability
This commit adds the automatic upgrading capability. It’s an
experimental feature that checks for the Kubescape Helm chart updates in
the background and upgrades the Kubescape release accordingly.

Since it comes with an elevated security risk, users should make sure to
read the associated README changes before they decide to use it.

Signed-off-by: Vlad Klokun <vklokun@protonmail.ch>